+++ src/link/batch-http/batchHttpLink.ts
@@ -25,7 +25,7 @@ export namespace BatchHttpLink {
   export type Options = Pick<
     BatchLink.Options,
     'batchMax' | 'batchDebounce' | 'batchInterval' | 'batchKey'
-  > & HttpOptions;
+  > & Omit<HttpOptions, 'useGETForQueries'>;
 }
 
 /**